# refactor: move segment generation into TableMutationAggregator for MERGE INTO