--- a/src/query/service/src/pipelines/builders/builder_replace_into.rs
+++ b/src/query/service/src/pipelines/builders/builder_replace_into.rs
@@ -49,7 +49,6 @@ use databend_common_sql::BindContext;
 use databend_common_sql::Metadata;
 use databend_common_sql::MetadataRef;
 use databend_common_sql::NameResolutionContext;
-use databend_common_storages_fuse::operations::new_serialize_segment_pipe_item;
 use databend_common_storages_fuse::operations::BroadcastProcessor;
 use databend_common_storages_fuse::operations::ReplaceIntoProcessor;
 use databend_common_storages_fuse::operations::TransformSerializeBlock;
@@ -134,13 +133,6 @@ impl PipelineBuilder {
         let mut block_builder = serialize_block_transform.get_block_builder();
         block_builder.source_schema = table.schema_with_stream();
 
-        let serialize_segment_transform = new_serialize_segment_pipe_item(
-            InputPort::create(),
-            OutputPort::create(),
-            table,
-            *block_thresholds,
-            replace.table_meta_timestamps,
-        )?;
         if !*need_insert {
             if segment_partition_num == 0 {
                 return Ok(());
@@ -204,15 +196,9 @@ impl PipelineBuilder {
         };
 
         // 4. connect with MergeIntoOperationAggregators
-        if segment_partition_num == 0 {
-            let dummy_item = create_dummy_item();
-            self.main_pipeline.add_pipe(Pipe::create(2, 2, vec![
-                serialize_segment_transform,
-                dummy_item,
-            ]));
-        } else {
+        if segment_partition_num != 0 {
             //      ┌──────────────────┐               ┌────────────────┐
-            // ────►│  SerializeBlock  ├──────────────►│SerializeSegment│
+            // ────►│  SerializeBlock  ├──────────────►│ DummyTransform │
             //      └──────────────────┘               └────────────────┘
             //
             //      ┌───────────────────┐              ┌──────────────────────┐
@@ -230,7 +216,7 @@ impl PipelineBuilder {
             let item_size = segment_partition_num + 1;
             let mut pipe_items = Vec::with_capacity(item_size);
             // setup the dummy transform
-            pipe_items.push(serialize_segment_transform);
+            pipe_items.push(create_dummy_item());
 
             let max_threads = self.settings.get_max_threads()?;
             let io_request_semaphore = Arc::new(Semaphore::new(max_threads as usize));