--- a/src/query/storages/fuse/src/operations/common/meta/mutation_log.rs
+++ b/src/query/storages/fuse/src/operations/common/meta/mutation_log.rs
@@ -39,7 +39,7 @@ pub enum MutationLogEntry {
         format_version: FormatVersion,
         summary: Statistics,
     },
-    ReclusterAppendBlock {
+    AppendBlock {
         block_meta: Arc<BlockMeta>,
     },
     DeletedBlock {