--- a/src/query/storages/fuse/src/operations/common/processors/mod.rs
+++ b/src/query/storages/fuse/src/operations/common/processors/mod.rs
@@ -26,6 +26,5 @@ pub use transform_block_writer::TransformBlockWriter;
 pub use transform_merge_commit_meta::TransformMergeCommitMeta;
 pub use transform_mutation_aggregator::TableMutationAggregator;
 pub use transform_serialize_block::TransformSerializeBlock;
-pub use transform_serialize_segment::new_serialize_segment_pipe_item;
 pub use transform_serialize_segment::new_serialize_segment_processor;
 pub use transform_serialize_segment::TransformSerializeSegment;