--- a/src/query/storages/fuse/src/operations/common/processors/transform_serialize_segment.rs
+++ b/src/query/storages/fuse/src/operations/common/processors/transform_serialize_segment.rs
@@ -27,7 +27,6 @@ use databend_common_pipeline_core::processors::InputPort;
 use databend_common_pipeline_core::processors::OutputPort;
 use databend_common_pipeline_core::processors::Processor;
 use databend_common_pipeline_core::processors::ProcessorPtr;
-use databend_common_pipeline_core::PipeItem;
 use databend_storages_common_table_meta::meta::column_oriented_segment::*;
 use databend_storages_common_table_meta::meta::BlockMeta;
 use databend_storages_common_table_meta::meta::SegmentInfo;
@@ -97,17 +96,6 @@ impl<B: SegmentBuilder> TransformSerializeSegment<B> {
             is_column_oriented: table.is_column_oriented(),
         }
     }
-
-    pub fn into_processor(self) -> Result<ProcessorPtr> {
-        Ok(ProcessorPtr::create(Box::new(self)))
-    }
-
-    pub fn into_pipe_item(self) -> PipeItem {
-        let input = self.input.clone();
-        let output = self.output.clone();
-        let processor_ptr = ProcessorPtr::create(Box::new(self));
-        PipeItem::create(processor_ptr, vec![input], vec![output])
-    }
 }
 
 pub fn new_serialize_segment_processor(
@@ -139,39 +127,6 @@ pub fn new_serialize_segment_processor(
                 table_meta_timestamps,
             );
             Ok(ProcessorPtr::create(Box::new(processor)))
-        }
-    }
-}
-
-pub fn new_serialize_segment_pipe_item(
-    input: Arc<InputPort>,
-    output: Arc<OutputPort>,
-    table: &FuseTable,
-    thresholds: BlockThresholds,
-    table_meta_timestamps: TableMetaTimestamps,
-) -> Result<PipeItem> {
-    match table.segment_format {
-        FuseSegmentFormat::Row => {
-            let processor = TransformSerializeSegment::new(
-                input,
-                output,
-                table,
-                thresholds,
-                RowOrientedSegmentBuilder::default(),
-                table_meta_timestamps,
-            );
-            Ok(processor.into_pipe_item())
-        }
-        FuseSegmentFormat::Column => {
-            let processor = TransformSerializeSegment::new(
-                input,
-                output,
-                table,
-                thresholds,
-                ColumnOrientedSegmentBuilder::new(table.schema(), thresholds.block_per_segment),
-                table_meta_timestamps,
-            );
-            Ok(processor.into_pipe_item())
         }
     }
 }