--- a/tests/sqllogictests/suites/base/09_fuse_engine/09_0026_merge_into.test
+++ b/tests/sqllogictests/suites/base/09_fuse_engine/09_0026_merge_into.test
@@ -2,9 +2,6 @@ statement ok
 set enable_distributed_merge_into = 1;
 
 statement ok
-set enable_experimental_merge_into = 1;
-
-statement ok
 drop database if exists db;
 
 statement ok
@@ -1402,9 +1399,6 @@ when matched then update *
 when not matched then insert *;
 ----
 0 40000
-
-statement ok
-set enable_experimental_merge_into = 0;
 
 statement ok
 set enable_distributed_merge_into = 0;