--- a/tests/sqllogictests/suites/mode/cluster/merge_into_non_equal_distributed.test
+++ b/tests/sqllogictests/suites/mode/cluster/merge_into_non_equal_distributed.test
@@ -1,7 +1,4 @@
 statement ok
-set enable_experimental_merge_into = 1;
-
-statement ok
 set enable_distributed_merge_into = 1;
 
 statement ok
@@ -297,9 +294,6 @@ select * from t1 order by a;
 6
 9
 10
-
-statement ok
-set enable_experimental_merge_into = 0;
 
 statement ok
 set enable_distributed_merge_into = 0;