--- a/tests/sqllogictests/suites/udf_server/udf_server_test.test
+++ b/tests/sqllogictests/suites/udf_server/udf_server_test.test
@@ -530,17 +530,11 @@ create or replace table target_udf(a string,b string,c string);
 statement ok
 create or replace table source_udf(a2 string,b2 string,c2 string);
 
-statement ok
-set enable_experimental_merge_into = 1;
-
 statement error 1065
 merge into target_udf using source_udf on target_udf.a = source_udf.a2 when not matched then insert (a) values(split_and_join(source_udf.a2, '; ', ':'));
 
 statement ok
 DROP FUNCTION split_and_join;
-
-statement ok
-set enable_experimental_merge_into = 0;
 
 ### enable udf for copy into table
 statement ok