… benchmark/clickbench/merge_into/queries.sql
… tests/suites/0_stateless/18_rbac/18_0001_udf_priv.sh