I hereby agree to the terms of the CLA available at: https://docs.databend.com/dev/policies/cla/

## Summary

This refactor improves the segment generation logic during MERGE INTO operations by centralizing it inside TableMutationAggregator. Previously, each node could produce multiple small segments, resulting in fragmented storage and inefficient compaction.

By delegating segment construction and aggregation to TableMutationAggregator, the system now accumulates and merges blocks more effectively, reducing the number of output segments and improving data layout.

Before:
```sql
mysql> create table t(a int , b int)row_per_block=10;
Query OK, 0 rows affected (0.14 sec)

mysql> insert into t select number, number%50 from numbers(100);
+-------------------------+
| number of rows inserted |
+-------------------------+
|                     100 |
+-------------------------+
1 row in set (0.36 sec)
Read 100 rows, 800.00 B in 0.265 sec., 377.59 rows/sec., 2.95 KiB/sec.

mysql> create table t1(a int , b int)row_per_block=10;
Query OK, 0 rows affected (0.14 sec)

mysql> insert into t1 select number, number % 25 from numbers(150) where number % 3 = 1;
+-------------------------+
| number of rows inserted |
+-------------------------+
|                      50 |
+-------------------------+
1 row in set (0.29 sec)
Read 150 rows, 1.17 KiB in 0.240 sec., 625.74 rows/sec., 4.89 KiB/sec.

mysql> merge into t using t1 on t1.a=t.a when matched then update set t.b=t1.b when not matched then insert *;
+-------------------------+------------------------+
| number of rows inserted | number of rows updated |
+-------------------------+------------------------+
|                      17 |                     33 |
+-------------------------+------------------------+
1 row in set (0.60 sec)
Read 150 rows, 840.00 B in 0.459 sec., 327.07 rows/sec., 1.79 KiB/sec.

mysql> select segment_count,block_count,row_count from fuse_snapshot('default','t') limit 3;
+---------------+-------------+-----------+
| segment_count | block_count | row_count |
+---------------+-------------+-----------+
|             2 |          15 |       117 |
|             4 |          17 |       117 |
|             1 |          10 |       100 |
+---------------+-------------+-----------+
3 rows in set (0.04 sec)
Read 3 rows, 609.00 B in 0.018 sec., 163.91 rows/sec., 32.49 KiB/sec.
```

After:
```sql
mysql> merge into t using t1 on t1.a=t.a when matched then update set t.b=t1.b when not matched then insert *;
+-------------------------+------------------------+
| number of rows inserted | number of rows updated |
+-------------------------+------------------------+
|                      17 |                     33 |
+-------------------------+------------------------+
1 row in set (0.78 sec)
Read 150 rows, 840.00 B in 0.577 sec., 259.85 rows/sec., 1.42 KiB/sec.

mysql> select segment_count,block_count,row_count from fuse_snapshot('default','t') limit 3;
+---------------+-------------+-----------+
| segment_count | block_count | row_count |
+---------------+-------------+-----------+
|             1 |          10 |       117 |
|             2 |          17 |       117 |
|             1 |          10 |       100 |
+---------------+-------------+-----------+
3 rows in set (0.07 sec)
Read 3 rows, 609.00 B in 0.032 sec., 94.95 rows/sec., 18.82 KiB/sec.
```

## Tests

- [ ] Unit Test
- [x] Logic Test
- [ ] Benchmark Test
- [ ] No Test - _Explain why_

## Type of change

- [ ] Bug Fix (non-breaking change which fixes an issue)
- [ ] New Feature (non-breaking change which adds functionality)
- [ ] Breaking Change (fix or feature that could cause existing functionality not to work as expected)
- [ ] Documentation Update
- [x] Refactoring
- [ ] Performance Improvement
- [ ] Other (please describe):

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/databendlabs/databend/17800)
<!-- Reviewable:end -->